requirements.txt

transformers
python-docx
PyPDF2
torch
